add context builder (#206)
