--- rclrs/src/builder.rs
@@ -1,3 +1,0 @@
-pub struct ContextBuilder {
-
-}
