+++ rclrs/src/context.rs
@@ -1,14 +1,11 @@
 mod builder;
 
 use crate::rcl_bindings::*;
-use crate::{RclrsError, ToResult};
+use crate::RclrsError;
 
 use self::builder::*;
-use std::ffi::CString;
-use std::os::raw::c_char;
 use std::string::String;
 use std::sync::Arc;
-use std::vec::Vec;
 
 use parking_lot::Mutex;
 
@@ -63,49 +60,8 @@ impl Context {
     /// assert!(Context::new(invalid_remapping).is_err());
     /// ```
     pub fn new(args: impl IntoIterator<Item = String>) -> Result<Self, RclrsError> {
-        // SAFETY: Getting a zero-initialized value is always safe
-        let mut rcl_context = unsafe { rcl_get_zero_initialized_context() };
-        let cstring_args: Vec<CString> = args
-            .into_iter()
-            .map(|arg| {
-                CString::new(arg.as_str()).map_err(|err| RclrsError::StringContainsNul {
-                    err,
-                    s: arg.clone(),
-                })
-            })
-            .collect::<Result<_, _>>()?;
-        // Vector of pointers into cstring_args
-        let c_args: Vec<*const c_char> = cstring_args.iter().map(|arg| arg.as_ptr()).collect();
-        unsafe {
-            // SAFETY: No preconditions for this function.
-            let allocator = rcutils_get_default_allocator();
-            // SAFETY: Getting a zero-initialized value is always safe.
-            let mut rcl_init_options = rcl_get_zero_initialized_init_options();
-            // SAFETY: Passing in a zero-initialized value is expected.
-            // In the case where this returns not ok, there's nothing to clean up.
-            rcl_init_options_init(&mut rcl_init_options, allocator).ok()?;
-            // SAFETY: This function does not store the ephemeral init_options and c_args
-            // pointers. Passing in a zero-initialized rcl_context is expected.
-            let ret = rcl_init(
-                c_args.len() as i32,
-                if c_args.is_empty() {
-                    std::ptr::null()
-                } else {
-                    c_args.as_ptr()
-                },
-                &rcl_init_options,
-                &mut rcl_context,
-            )
-            .ok();
-            // SAFETY: It's safe to pass in an initialized object.
-            // Early return will not leak memory, because this is the last fini function.
-            rcl_init_options_fini(&mut rcl_init_options).ok()?;
-            // Move the check after the last fini()
-            ret?;
-        }
-        Ok(Self {
-            rcl_context_mtx: Arc::new(Mutex::new(rcl_context)),
-        })
+        let builder = ContextBuilder::new(args)?;
+        builder.build()
     }
 
     /// Checks if the context is still valid.
