+++ rclrs/src/lib.rs
@@ -5,26 +5,20 @@
 //!
 //! [1]: https://github.com/ros2-rust/ros2_rust/blob/main/README.md
 
-mod clock;
 mod context;
-mod duration;
 mod error;
 mod node;
 mod parameter;
 mod qos;
-mod time;
 mod wait;
 
 mod rcl_bindings;
 
-pub use clock::*;
 pub use context::*;
-pub use duration::*;
 pub use error::*;
 pub use node::*;
 pub use parameter::*;
 pub use qos::*;
-pub use time::*;
 pub use wait::*;
 
 use rcl_bindings::rcl_context_is_valid;
